% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galaxies.R
\docType{data}
\name{galaxies}
\alias{galaxies}
\title{Velocities and distances for 24 galaxies in relation to Earth.}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{Galaxy}{unique id, NGC/IC}
\item{velocity}{velocity away from Earth, in km per second}
\item{distance}{distance from Earth, in megaparsecs (3.086 x 10^19 km)}
}
}
\source{
Peter Bajorski, RIT 2021
}
\usage{
galaxies
}
\description{
A selection of 24 galaxies from the \emph{New General Catalogue of Nebulae and
Clusters of Stars} (NGC) and the supplemental \emph{Index Catalogues} (IC). Data
provided by course instructor for STAT 745.
}
\keyword{datasets}
